requirements.txt
asgiref==3.8.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.44.1
cryptography==42.0.7
defusedxml==0.8.0rc2
Django==5.0.6
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-filter==24.2
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-nested-routers==0.94.1
gunicorn==23.0.0
idna==3.7
Markdown==3.6
mysqlclient==2.2.4
oauthlib==3.2.2
packaging==25.0
pillow==10.3.0
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.1.1
python3-openid==3.2.0
razorpay==1.4.2
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==80.9.0
six==1.17.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.9.0
